// src/types/Slug.ts
export enum VALKYRIES {
  AZURE_EMPYREA = 'ae',
  BLOOD_ROSE = 'br',
  BRIGHT_KNIGHT_EXCELSIS = 'bke',
  DIVINE_PRAYER = 'dp',
  INFINITE_OUROBOROS = 'io',
  LIGHTNING_EMPRESS = 'le',
  SNOWY_SNIPER = 'ss',
  VALKYRIE_CHARIOT = 'vc',
  VERMILION_KNIGHT_ECLIPSE = 'vke',
  SILVERWING_NEX = 'sw',
  ARGENT_KNIGHT_ARTEMIS = 'aka',
  HERRSCHER_OF_FLAMESCION = 'hofs',
  DEA_ANCHORA = 'da',
  STARCHASM_NYX = 'nyx',
  HERRSCHER_OF_SENTIENCE = 'hos',
  HERRSCHER_OF_REASON = 'hor',
  PRINZESSIN_DER_VERURTEILUNG = 'pv',
  HERRSCHER_OF_THUNDER = 'hot',
  MIDNIGHT_ABSINTHE = 'rvn',
  MISS_PINK_ELF = 'ely',
  TWILIGHT_PALADIN = 'tp',
  STRIKER_FULMINATA = 'vb',
  STYGIAN_NYMPH = 'sn',
  VALKYRIE_GLORIA = 'vg',
  RITUAL_IMAYOH = 'ri',
  LUNA_KINDRED = 'lk',
  SWEET_N_SPICY = 'sns',
}

export const FLAME_CHASERS = {
  ELYSIA: 'elysia',
};

export const SIGNET_GROUP = {
  ELYSIA_HOT: `${FLAME_CHASERS.ELYSIA}-${VALKYRIES.HERRSCHER_OF_THUNDER}`,
};

export const SIGNETS = {
  ELYSIA_HOT_1: `${SIGNET_GROUP.ELYSIA_HOT}-1`,
};